Fix server side large message receiving
The check for whether to request more data was all messed up.
On the client this needs to be after initial metadata is sent to guarantee that we have a live stream, but there's no such requirement at the server. Fix this.

I also ran clang-format across the code so there are some formatting changes.