Avoid identifiers reserved by Ruby
"prefix `rb` shows that it belongs to `ruby`". [RHG]

`rb` is a prefix to prevent exported symbols in libruby from
conflicting with symbols in other libraries, in particular,
dynamically loaded Ruby extension libraries.
So the prefix is reserved by Ruby by convention.

For gRPC, simply "grpc_" would be a better prefix.

[RHG]: Minero Aoki. "The Ruby Hacking Guide" (2002), chapter 4