Fix accelerated wakeups
We can end up in situations where a pollset needs to kick itself.

This is supposed to be an accelerated codepath, however a bug crept in
whereby we missed the opportunity to do so, resulting in needing to
round trip these wakeups redundantly through the OS and wake other
threads unnecessarily.