Rename Fail Fast doc to Wait for Ready
The intention is to deminish the term "fail fast" and instead encourage
"wait for ready." However, the semantics of "the opposite of fail fast"
weren't really documented since that was the previous default, so it
became a bit of a rewrite to define both terms.

The file name uses dashes instead of underscore, since dashes are more
common in the docs.