Correct Python cancel_after_begin interop test
It was a mistake that requests might be sent; the test specification
calls for no requests to be sent. It was a mistake that the response
future's cancelled() method was called; the cancelled() method returns
something more like "was this object's cancel() method called earlier?"
than "did the RPC terminate with status code CANCELLED?". Since it's
something that we'd well enough like to work I've retained the
cancelled() call with a different failure message.