Add a spinlock type.
Useful for situations where we need to repeatedly trylock, not useful
for cases where we need to lock (due to spinning).

Add a variant of sync_test to test it (with the same tests we run for
gpr_mu).

Add a benchmark to bm_closure to demonstrate single threaded performance.