Fix for using gem build/install
When pre-compiled assets are missing for the current platform, and you run gem install, the compiled native extensions are placed in the first require paths folder. This differs from rake-compiler task, where they are always placed in lib.

Setting the lib path as the first require path will ensure the compiled library is placed in the correct folder.